pyproject.toml
[project]
name = "time-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "websockets>=15.0.1",
]
